Fix TizenConnectionHandler.kt SAAgent misuse
